+++ providers/offline/src/providers/paragraph.offline.provider.ts
@@ -1,7 +1,7 @@
 import type {Paragraph, ParagraphData} from '@deckdeckgo/editor';
-import {del, get, set} from 'idb-keyval';
+import {get, set, update} from 'idb-keyval';
 
-export const createOfflineParagraph = ({
+export const createOfflineParagraph = async ({
   docId,
   paragraphData,
   paragraphId
@@ -10,58 +10,36 @@ export const createOfflineParagraph = ({
   paragraphData: ParagraphData;
   paragraphId: string;
 }): Promise<Paragraph> => {
-  return new Promise<Paragraph>(async (resolve, reject) => {
-    try {
-      const now: Date = new Date();
+  const now: Date = new Date();
 
-      const paragraph: Paragraph = {
-        id: paragraphId,
-        data: {
-          ...paragraphData,
-          created_at: now,
-          updated_at: now
-        }
-      };
-
-      await set(`/docs/${docId}/paragraphs/${paragraph.id}`, paragraph);
-
-      resolve(paragraph);
-    } catch (err) {
-      reject(err);
+  const paragraph: Paragraph = {
+    id: paragraphId,
+    data: {
+      ...paragraphData,
+      created_at: now,
+      updated_at: now
     }
-  });
+  };
+
+  await set(`/docs/${docId}/paragraphs/${paragraph.id}`, paragraph);
+
+  return paragraph;
 };
 
 export const getOfflineParagraph = ({docId, paragraphId}: {docId: string; paragraphId: string}): Promise<Paragraph | undefined> =>
   get(`/docs/${docId}/paragraphs/${paragraphId}`);
 
-export const updateOfflineParagraph = ({docId, paragraph}: {docId: string; paragraph: Paragraph}): Promise<Paragraph> => {
-  return new Promise<Paragraph>(async (resolve, reject) => {
-    try {
-      if (!paragraph || !paragraph.data) {
-        reject('Invalid paragraph data');
-        return;
-      }
-
-      paragraph.data.updated_at = new Date();
-
-      await set(`/docs/${docId}/paragraphs/${paragraph.id}`, paragraph);
-
-      resolve(paragraph);
-    } catch (err) {
-      reject(err);
+export const updateOfflineParagraph = ({docId, paragraph}: {docId: string; paragraph: Paragraph}): Promise<void> =>
+  update(`/docs/${docId}/paragraphs/${paragraph.id}`, ({data, ...rest}: Paragraph) => ({
+    ...rest,
+    data: {
+      ...data,
+      ...paragraph.data,
+      updated_at: new Date()
     }
-  });
-};
+  }));
 
-export const deleteOfflineParagraph = ({docId, paragraphId}: {docId: string; paragraphId: string}): Promise<void> => {
-  return new Promise<void>(async (resolve, reject) => {
-    try {
-      await del(`/docs/${docId}/paragraphs/${paragraphId}`);
-
-      resolve();
-    } catch (err) {
-      reject(err);
-    }
-  });
-};
+export const deleteOfflineParagraph = ({docId, paragraphId}: {docId: string; paragraphId: string}): Promise<void> =>
+  update(`/docs/${docId}/paragraphs/${paragraphId}`, ({data, ...rest}: Paragraph) => ({
+    ...rest
+  }));