+++ studio/src/global/app-staging.ts
@@ -68,10 +68,10 @@ setupDeckGoConfig({
   },
   cloud: {
     api: {
-      cdn: 'https://unpkg.com/@deckdeckgo/firebase@4.1.3/dist/deckdeckgo-firebase/index.esm.js'
+      cdn: 'https://unpkg.com/@deckdeckgo/firebase@4.2.0/dist/deckdeckgo-firebase/index.esm.js'
     },
     signIn: {
-      cdn: 'https://unpkg.com/@deckdeckgo/firebase@4.1.3/dist/deckdeckgo-firebase/deckdeckgo-firebase.esm.js',
+      cdn: 'https://unpkg.com/@deckdeckgo/firebase@4.2.0/dist/deckdeckgo-firebase/deckdeckgo-firebase.esm.js',
       tag: 'deckgo-firebase-signin'
     }
   },