+++ utils/editor/src/index.ts
@@ -10,6 +10,7 @@ export * from './models/data/slide';
 export * from './models/data/template';
 export * from './models/data/token';
 export * from './models/data/user';
+export * from './models/data/data';
 
 export * from './models/storage/storage';