+++ providers/offline/CHANGELOG.md
@@ -1,3 +1,17 @@
+# 0.0.7 (2022-08-15)
+
+### Breaking Changes
+
+- `delete` does not delete the all idb entry but `update` and remove `data` object
+
+### Features
+
+- atomic `update`
+
+### Build
+
+- bump dependencies
+
 # 0.0.5 - 0.0.6 (2022-02-25)
 
 ### Features