+++ webcomponents/studio/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.0.22 (2022-08-15)
+
+### Build
+
+- bump dependencies
+
 # 0.0.21 (2022-05-26)
 
 ### Features